% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in compute_results.R
\name{compute_results}
\alias{compute_results}
\title{Compute the results}
\usage{
compute_results(sol_bln, sol_cfl)
}

\arguments{
\item{sol_bln}{a baseline solution.}

\item{sol_cfl}{a counterfactual solution.}
}

\description{
Compute the results
}

\keyword{internal}
