% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cp2015-package.R
\name{cp2015nafta}
\alias{cp2015nafta}
\title{Data to Replicate Section 5.1}
\description{
To quantify the trade and welfare effects of NAFTA. Welfare
effects from NAFTA's tariff reductions while fixing the tariff to and from
the rest of the world to the year 1993.
}

