% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in prepare_cp2015.R
\name{prepare_cp2015}
\alias{prepare_cp2015}
\title{Prepares the components for the Caliendo and Parro model (2015)}
\usage{
prepare_cp2015(data, zero_aggregate_deficit = FALSE)
}

\arguments{
\item{data}{a named list with the data to create the model components.
Run \code{help(cp2015nafta)} to see the data format that is required.}

\item{zero_aggregate_deficit}{whether to impose zero aggregate deficits.}
}

\value{
a named list with 5 elements.
}

\description{
Given a database, it builds the elements needed to run the simulations.
}

\keyword{internal}
