% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cp2015.R
\name{prepare_cp2015}
\alias{prepare_cp2015}
\title{Prepares the components for the Caliendo and Parro model (2015)}
\usage{
prepare_cp2015(data)
}
\arguments{
\item{data}{a named list with the basedata to create the model components.
Run \code{help(cp2015_nafta)} to see the data format that is required.}
}
\value{
a named list with 5 elements.
}
\description{
Given a database, it builds the elements needed to run the simulations.
}
\details{
The returned elements in the list are:
\itemize{
\item parameters. Elements that do not change during simulation.
It does not include exogenous variables as the user may change these variables
in the scenario specifications.
\item variables. All model variables. Endogenous and Exogenous.
\item sets. Vectors that list the regions, sectors and production factors
present in the data.
\item parameters_descs. Coefficient descriptions.
\item variables_descs. Variable descriptions.
}
}
\keyword{internal}
