% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in run_cp2015.R
\name{run_cp2015}
\alias{run_cp2015}
\title{Run a simulation using the Caliendo and Parro (2015) quantitative trade model}
\usage{
run_cp2015(
  data,
  zero_aggregate_deficit = FALSE,
  ufactor = 0.5,
  tol = 1e-07,
  maxiter = 10000,
  verbose = TRUE,
  triter = 100
)
}

\arguments{
\item{data}{a List with the model data. Run \code{help(cp2015nafta)} to see
the required data.}

\item{zero_aggregate_deficit}{a boolean indicating whether the simulation
should impose zero aggregate deficits.}

\item{ufactor}{an update factor between (0, 1]. This value is used
to update the value of variables at each iteration.}

\item{tol}{a tolerance number for convergence. The default is 1e-7.}

\item{maxiter}{the number of maximum iterations.}

\item{verbose}{a boolean indicating whether convergence information should be printed.}

\item{triter}{an integer indicating that information should be printed for each
multiple of that number.}
}

\value{
A list with 13 elements:

 * c_nj_hat (changes in cost an input bundle) - a data.table with 3 columns:
   * region
   * sector
   * c_nj_hat (relative change)
 * P_nj_hat (changes in the region-sector price index) - a data.table with 3 columns:
   * region
   * sector
   * P_nj_hat (relative change)
 * pi_nij (bilateral trade share) - a data.table with 5 columns:
   * importer
   * exporter
   * sector
   * pi_bln (trade share in the baseline scenario)
   * pi_cfl (trade share in the counterfactual scenario)
 * X_nj (total expenditure) - a data.table with 4 columns:
   * region
   * sector
   * X_bln (expenditure in the baseline scenario)
   * X_cfl (expenditure in the counterfactual scenario)
 * I_n (regional income) - a data.table with 3 columns:
   * region
   * I_bln (regional income in the baseline scenario)
   * I_cfl (regional income in the counterfactual scenario)
 * P_n_hat (consumer price index) - a data.table with 2 columns:
   * region
   * P_n_hat (relative change)
 * w_n_hat ("wages") - a data.table with 2 columns:
   * region
   * w_hat (relative change)
 * trade (trade data) - a data.table with 9 columns:
   * importer
   * exporter
   * sector
   * tau_bln (tariffs in the baseline scenario)
   * tau_cfl (tariffs in the counterfactual scenario)
   * d_bln (relative changes of the iceberg trade costs in the baseline scenario)
   * d_cfl (relative changes of the iceberg trade costs in the counterfactual scenario)
   * trade_bln (trade value, net of tariffs, in the baseline scenario)
   * trade_cfl (trade value, net of tariffs, in the counterfactual scenario)
 * tot (terms of trade) - a data.table with 4 columns:
   * partner
   * region
   * sector
   * tot (contribution of terms of trade to welfare in \%)
 * vot (volume of trade) - a data.table with 4 columns:
   * region
   * partner
   * sector
   * vot (contribution of volume of trade to welfare in \%)
 * tech (technical efficiency) - a data.table with 4 columns:
   * region
   * partner
   * sector
   * tech (contribution of technical efficiency to welfare in \%)
 * welfare (total welfare by region) - a data.table with 6 columns:
   * region
   * total (total welfare in \%)
   * tot (contribution of terms of trade to welfare in \%)
   * vot (contribution of volume of trade to welfare in \%)
   * realwage (relative change in the real wage).
   * tech (contribution of technical efficiency to welfare in \%)
 * convergence_info (info about the solution) - a data.table with 3 variables:
   * scenario
   * criteria_value
   * message
}

\description{
Caliendo and Parro (2015) develop a Ricardian quantitative trade model that
considers multiple countries and multiple sectors. The model allows interactions
across sectors based on Input-Output linkages.

The model is originally used to study the trade and welfare effects of NAFTA
given the observed tariff changes. However, from data provided by the user, this model
can be used for simulations to assess the effects of different trade
policies (changes in tariffs and/or iceberg trade costs).
}

\examples{
\dontrun{
data("cp2015nafta")

results_without_deficits <- run_cp2015(
  data = cp2015nafta,
  zero_aggregate_deficit = TRUE,
  verbose = TRUE
)

results_with_deficits <- run_cp2015(
  data = cp2015nafta,
  zero_aggregate_deficit = FALSE,
  verbose = TRUE
)
}
}

