% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cp2015.R
\name{run_cp2015}
\alias{run_cp2015}
\title{Run a simulation using the Caliendo and Parro (2015) model}
\usage{
run_cp2015(
  data,
  zero_aggregate_deficit = FALSE,
  ufactor = 0.5,
  tol = 1e-07,
  maxiter = 10000,
  verbose = TRUE,
  triter = 100,
  nthreads = 1
)
}
\arguments{
\item{data}{a List with the model data. Run \code{help(cp2015_nafta)} to see
the required data.}

\item{zero_aggregate_deficit}{a boolean indicating whether the simulation
should impose zero aggregate deficits.}

\item{ufactor}{a double with an update factor number between (0, 1]. This value is used
to update the value of variables at each iteration.}

\item{tol}{a double with the tolerance criteria.}

\item{maxiter}{an integer with the number of maximum iterations.}

\item{verbose}{a boolean indicating whether convergence information should be printed.}

\item{triter}{an integer indicating that information should be printed for each
multiple of that number.}

\item{nthreads}{an integer indicating the number of threads to use.}
}
\description{
Run a simulation using the Caliendo and Parro (2015) model
}
