% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{solve_model}
\alias{solve_model}
\title{Solve model}
\usage{
solve_model(data, ufactor, tol, maxiter, triter, trace)
}

\arguments{
\item{data}{a List with the model data.}

\item{ufactor}{a double with an update factor number between (0, 1]. This value
  is used to update the value of variables at each iteration.}

\item{tol}{a double with the tolerance criteria.}

\item{maxiter}{an integer with the number of maximum iterations.}

\item{triter}{an integer indicating that information should be printed for each
  multiple of that number.}

\item{trace}{a boolean indicating whether convergence information should be
  printed.}
}

\description{
Solve model
}

